Add slug to database converter
This converts slugs that are longer than 63 characters so that the
underlying database name is 63 characters long. It does this by taking a
hash of the original slug, then appending it to a shortened version of
the slug.

This is necessary because of MongoDB's limit on the length of a database
name (which appears to be 64 characters, though the nodejs mongo module
only allows 63 characters).