Refactor the storage selector into a module
A big chunk of the logic in app.js was the storage selectors. This
commit moves that logic into a separate file and changes it so the
selector is exposed as a simple piece of middleware which the main app
doesn't have to care about other than passing it the options.