Make option names and docs clearer for hidden fields
As discussed in the Pull Request comments, using the hidden option isn't
the clearest way to reveal out intentions. I've changed hidden to
`fieldSpec` as that better describes what we're actually doing with the
config. I've also updated the README text to be more explicit about what
the settings in that field do.