Switch to elasticsearch for search
The double-metaphone search implementation was rather slow for longer
queries, it is also rather brittle as it relies on the document being
indexed in a certain way for it to be searchable. Switching to
elasticsearch means that the indexing can be done by simply handing it
the whole document, and from some initial testing it seems to be
returning more relevant results than double metaphone in mongo.

Elasticsearch does support double metaphone, though I believe you need a
plugin, so we can always revisit this approach in the future if needed.