Add module for reindexing an instance into elasticsearch
Takes a database name, connects to that database and then indexes the
documents from each popolo collection into elasticsearch. If a
document hasn't been indexed before it will be created, otherwise it
will be updated based on its _id field.

This also exposes the method on the main module so the popit UI can
require it and re-index all of the instances at once.