Make sure memberships are populated in the API
This is quite a crude initial implementation, but it does achieve the
desired outcome of populating the memberships array.

One worry about this is that it might give people the impression that
you can create memberships by POSTing them as part of the person, but I
guess thorough documentation can get around issues like this.