Move json transform into a plugin
This means that all of the add-on functionality for the core popolo
models exists in plugins, which hopefully makes the actual model
definitions a bit clearer.

I think further improvements can be made to the clarity of the model
definitions by merging the collections.js module into models.js. That
way there is one central place which defines the popolo models for the
application.