Add pagination for API search results
Rather than expose the whole elasticsearch uri search API [1] to the
user, we just allow them to specify the `q` parameter, which still
offers a rich query API, then layer our own pagination on top.

This gives an example of what the pagination query string API might
look like for other resources when we get round to doing #8. The
parameters are called `per_page` and `page` to allow specifying the
number of results to return and the page of results to return.

[1] http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-uri-request.html