Add member property to membership schema
I'm adding this property to the popolo+mysociety/membership.json schema
initially, as this isn't officially part of popolo yet.

This property will be used as a polymorphic reference to the person OR
organization that is related to the organization of the membership.

This is to help support the concept of coalitions in PopIt, the idea being
that there is an organization which represents the coalition, then the various
political parties that are part of the coalition have a membership
record where the party is the "member" of the membership, and the
coalition itself is referenced through the "organization" property.