Add next/prev links and has_more flag to pagination metadata
Adds next_url/prev_url properties to the API output if there are more
pages of results. I've also added a has_more boolean flag which
indicates if there are more pages of results, although this can be
detected by the absence of a next_url property, it seems nicer to have
it in here explicitly.