Filter collections by ?at parameter
This allows filtering a whole collection using the ?at parameter to
remove entries which wouldn't have existed on a given date.

This isn't fully working but I'm wondering if we want/need to do this at
all. The same query can actually be achieved using the search
interface, so I'm unsure if this code should go in at all.

Currently this only works for records with both a start and an end date.