Allow the default language to be configured
Allow the user to specify a default language to fall back to when none
of the requested ones match. This can be configured when popit-api is
run standalone, but can also be used when the app is embedded in popit,
which will allow the user to select a default language for their
instance.