Fix membership indexing issues
When indexing a large number of memberships the process was getting
killed with an out of memory error. This is because for each membership
the related person, post and organization was being looked up, then in turn
for each related person, post or organization their memberships were being
looked up.

This change skips indexing person, post and organization memberships
when embedding them in an indexed membership.