Change findMemberships to be an instance method
Rather than having this as a private implementation we expose the method
on the instance so we can actively find memberships when we need then.

This also has the added bonus of making the method signature simpler.
When we get rid of the pre init hook we can also drop the id argument
and just use the instance's id.