fix elastic search date mapping and search issues
Elastic search isn't happy with saving blank dates so we put in some
place holder ones and mark them as such when indexing documents. We then
remove the place holders and markers when we pull the dates back out.

This also means that death_date:>$some_date etc works as expected when
using the search.

FIxes #63 and Fixes mysociety/popit#610