When PUTing a record replace the whole document
This is breaking behaviour, previously only the fields included in the
request would be updated, but with this change the whole document gets
replaced with the request body.

This is probably fine most of the time, but there will inevitably be
someone who depends on the old behaviour, so this would be a backwards
incompatible change.