create memberships included inline when POSTing
If a document POSTed includes a memberships field with memberships in
then these will be created along with the document. If there's a problem
then nothing will be created and an error returned.

Automatically fills in the relevant id in the membership if it's missing
and also falls over if the id in the membership does not match the
created document.

Returns the documument with the memberships included if successful.