include link to image proxy version with images
If the proxyBaseUrl option is set then we include a link as a proxy_url
key for each image to the base proxy url of images so CPI clients can
link to thumbnailed versions. It doesn't include any resizing parameters
which the clients will need to add themselves.

Fixes #mysociety/popit/707