Create/update memberships included inline in documents
This changes the API to process memberships included in POSTs and PUTs
which create or update persons, organizations or posts. If a membership
key is included in these then the memberships of the document will be
updated to match this. It will create, update and delete memberships as
required in order to do this.

If no memberships key is present in the document them memberships will
be unchanged.

This also adds an API v1.0.0 as this change is backwards incompatible