Preserve ?embed and ?q parameters when paginating
When paginating we don't want to keep all parameters because it can
cause issue with jQuery's JSONP functionality (see
186c09bf738e4299bf4386864e3cd02e4cc6b7b5).

We do however want to keep user supplied parameters such as how many
results per_page, what to embed and the q parameter for the current
search.

This ensures that these parameters are kept and makes it easy to add
other parameters to keep in the future.