Remove i18n test that fails on travis
This test relies on a brittle setTimeout call to try and wait for
elasticsearch to reindex a document before it proceeds, however this
almost never works on travis, which leads to the tests being broken most
of the time.

Once the next version of mongoose is released it seems that async post
hooks on models will be possible, so at that point we can make the
reindexing happen before we return from the save call which will
sidestep the need to use a setTimeout.

Fixes #117