add name resolution end point to API
/:collection/resolve uses the resolve elasticsearch index to produce a
list of possible matches sorted by score. It searches on the alt_names
and the other_names for matches and then filters the results to give
one result per document with the highest score for that document attached.

Optionally it can use the organization id and a date the person is
expected to be alive on.

It also filters out the added name variations, added date information
and so on.