use nested filtered query for name resolution
In combination with telling elastic search to index memberships as a
nested type this means we can make sure to say we want a membership
apply on a certain date.

This does mean converting to using the hash version of the query rather
than the simple string version which significantly increases the
complexity of the code but this then allows us to fine tune the
query a bit more.

Also use phrase type of the match so it doesn't try and do clever
splitting down of the search name. This might not be a good idea.