Guess the MIME type of an uploaded image if it wasn't specified
If mime_type isn't set on an image object stored by PopIt then it
will currently be returned with "Content-Type: undefined"; to avoid
this, which the image proxy won't cope with.

This commit introduces code that infers the file's MIME type using
mmmagic (a node wrapper for libmagic) and sets mime_type on the image
object to that value.