Filter membership legislatures when ?at is present
When an ?at parameter is provided and the user has chosen to embed a
membership's legislature we want to filter out any memberships with
legislatures which aren't valid for the specified date.

Note that currently this only works for the v1.0.0-alpha version of the
API and only works when legislatures are embedded.